pyproject.toml
[tool.poetry]
name = "pingpong-slackbot"
version = "0.1.0"
description = ""
authors = ["Erlend Vollset <erlend.vollset@cognite.com>"]

[tool.black]
line-length = 120
target_version = ['py38']
include = '\.py$'

[tool.isort]
line_length=120                # corresponds to -w  flag
multi_line_output=3            # corresponds to -m  flag
include_trailing_comma=true    # corresponds to -tc flag
skip_glob = '^((?!py$).)*$'    # isort all Python files
float_to_top=true

[tool.pytest.ini_options]
addopts = "--doctest-modules --ignore=bulk-import.py"

[tool.poetry.dependencies]
python = "^3.10"
slack-sdk = "^3.11.2"
cognite-sdk-core = "^2.32.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.15.0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
